huge update. introducing interplanetary transfer. steering is not exact enough at the moment though :/

+++ Code/KOSM.Tasks/Time/WaitTask.cs
@@ -13,7 +13,7 @@ namespace KOSM.Tasks
         protected double secondsToWait = 0;
         protected double timeToWaitFor = -1;
 
-        public WaitTask(IWorld world, int secondsToWait)
+        public WaitTask(IWorld world, double secondsToWait)
         {
             this.secondsToWait = secondsToWait;
         }
